GUI Work
Made a class for the GUI elements
