+++ .gitignore
@@ -205,3 +205,4 @@ __marimo__/
 .streamlit/secrets.toml
 /__pycache__
 *.pyc
+*.pyc
